more explanation and test cases

+++ src/MM_UnitTestFramework/MM_UnitTestFramework.h
@@ -110,7 +110,7 @@ Output:
 	template<typename FirstType, typename... RestTypes>
 	void printDifferentTypesOfVariables(size_t width, string str, FirstType t, RestTypes... args)  // used when at least one variable is provided
 	{
-		int firstCommaPosition = str.find_first_of(',');
+		size_t firstCommaPosition = str.find_first_of(',');
 		printDifferentTypesOfVariables(width, str.substr(0, firstCommaPosition), t);
 		//MM_COUT << "\n";
 		printDifferentTypesOfVariables(width, str.substr(firstCommaPosition + 1, string::npos), args...);
